requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-cartesia
livekit-plugins-deepgram
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn